adds upgraded review modal
-replaces hardcode placeholders with dummy data
-fixes issue with router paths
