+++ car-reviews/src/components/MainPage/maincontent.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import placeholder from '../../logo.svg';
 import { Button } from 'reactstrap';
+import ReviewModal from '../Modals/reviewmodal';
 
 // This component generates Review and Reviewer cards. I chose to make the cards using buttons
 // because they will need to be clicked on to open the review page. This is rendered in MainPage.
@@ -10,13 +11,7 @@ class MainContent extends Component {
         return ( 
             <div className="main-content-border">
                 <h3>Featured Reviews</h3>
-                <Button className="main-card"> 
-                    <img src={placeholder} style={{ height: '60px', width: '60px' }} />
-                    <p>Star Rating</p> 
-                    <p>Year, Make, Model</p>
-                    <p>Trim</p>
-                    <p>Reviewer</p>
-                </Button>
+                <ReviewModal />
                 <h3>Popular Cars</h3>
                 <Button className="main-card"> 
                     <img src={placeholder} style={{ height: '60px', width: '60px' }} />
